chore: Transform anchor test to typescript (#23499)
* chore: Transform anchor test to typescript

* Separate dom

* use hash

* .

* chagnelog

* chore: Transform anchor test to typescript

* Separate dom

* use hash

* .

* rebase
